Fix syntax error w/ from-scratch-init playbook
Fix a syntax error which I didn't notice because I never actually tested
trying to run the playbook after the changes to not require provisioning
file systems.

At this point, my three Rasp Pi's are provisioned.

Next step is writing the base role and applying it to all of them. Then
we can think about application specific tasks.

For now, I hard code the IPs in the hosts file. They are not static IPs,
so could change if the RaspPis restart. However, that's fairly unlikely.

I'll look into getting static IPs when back in a more permanent
location.