Set up nginx-ssl for home-assistant
One important component is that we need to update our nginx config to
support forwarding web sockets. See
https://nginx.org/en/docs/http/websocket.html for details.

Additionally, we currently leave copying the `mattjmcnaughton.io` ssl
certs as a separate role/playbook. This makes sense because the
playbooks will be run at a different cadence and require different
inputs. However, we will need to ensure we run them together the first
time we create a new home-assistant machine, or else the newly
provisioned machine will only have the self-signed certs.

One fairly large smell here is that we copied the
`mattjmcnaughton.nginx` repo directly from `nuage`. We then made small
modifications to support web sockets, but they are essentially the same.
In the medium term, both `self-hosted` and `nuage` need to be using the
same roles. We will look into extracting common roles from the repo and
storing them on Github/ansible-galaxy.

For now, we just update `/etc/hosts` on my machines to point to the proper
ip, but we should have better DNS solution/mechanism for provisioning
static ips going forward.