Configure Mqtt integration via HA configuration.yaml
Update `configuration.yaml` for HA to configure MQTT (preventing us from
needing to configure via the UI).

However, I could not find a similar method for configuration the Tasmota
integration via the UI. For now, we will need to continue doing that via
the UI. I updated the docs for cloudfree-smart-plug2 accordingly.